requirements.txt
discord.py
git+https://github.com/Rapptz/discord-ext-menus
hentai
nekos.py